backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
httpx==0.25.0
yt-dlp==2023.10.13
openai==1.2.0
git+https://github.com/openai/whisper.git
python-multipart==0.0.6
websockets==11.0.3
ffmpeg-python==0.2.0
